+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -150,6 +150,20 @@ public class DisabledHub : IHub, IDisposable
     /// <summary>
     /// No-Op.
     /// </summary>
+    public void CaptureFeedback(SentryFeedback feedback, Action<Scope> configureScope)
+    {
+    }
+
+    /// <summary>
+    /// No-Op.
+    /// </summary>
+    public void CaptureFeedback(SentryFeedback feedback, SentryHint? hint, Action<Scope> configureScope)
+    {
+    }
+
+    /// <summary>
+    /// No-Op.
+    /// </summary>
     public void CaptureFeedback(SentryFeedback feedback, Scope? scope = null, SentryHint? hint = null)
     {
     }